Add: src/apis/passwordUpdates.jsをTypeScript対応させた
